refactor(cli): extract Supabase path validation into separate module
Move path validation logic to a new supabase-path-prompt module and simplify install command
implementation
